update signup page

+++ loginframe/src/main/java/com/example/loginframe/Repository/UserRepository.java
@@ -2,11 +2,9 @@ package com.example.loginframe.Repository;
 
 import com.example.loginframe.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
     Optional<User> findByEmail(String email);
